+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+### Fixed
+
+- Assigning to getters using the legacy decorator
+
 ## [1.2.0] - 2019-04-10
 
 ### Changed